Refactor lamp management and add graphical features
Refactored lamp addition to prevent multiple lamps in the same room and updated exception handling accordingly. Introduced DisegnoLampadina for graphical representation of lamps, updated CasaGrafica and Lampadina to use it, and added menu selection and deselection features. Improved Main to support interactive lamp selection and modification, and moved position data to a shared file. Various type and method signature adjustments for consistency and extensibility.

+++ CasaDomotica/src/CasaIntelligente.java
@@ -55,7 +55,7 @@ public class CasaIntelligente {
         gestore.spegniTutte();
     }
 
-    public void addLampadina(int potenza, Posizione pos, String nome, int intensita, String colore) throws LampadinaDuplicataException {
+    public void addLampadina(double potenza, Posizione pos, String nome, int intensita, String colore) throws LampadinaDuplicataException {
         Lampadina l = new Lampadina(gestore.newId(), potenza, pos, nome, intensita, colore);
         gestore.aggiungiLampadina(l);
     }
